[Divide by 0 Error] svd check migrate to c++, move svd test_0_size
