+++ python/paddle/tensor/linalg.py
@@ -1922,9 +1922,6 @@ def svd(x, full_matrices=False, name=None):
             #                  V * VH == I
     """
 
-    if x.size == 0:
-        raise ValueError("input size should not be 0")
-
     if in_dygraph_mode():
         return _C_ops.svd(x, full_matrices)
     else:
